+++ MercadoPagoSDK/MercadoPagoSDK/Core/Checkout/ViewModel/MercadoPagoCheckoutViewModel+NextStep.swift
@@ -230,7 +230,6 @@ extension MercadoPagoCheckoutViewModel {
     }
 
     func needOneTapFlow() -> Bool {
-        return false
         guard let search = self.search else {
             return false
         }
